+++ src/timing.c
@@ -1,11 +1,12 @@
 #include "timing.h"
 
+#include <stdint.h>
 #include <stdio.h>
 #include <time.h>
 
 void bellman_fn_execution_time(t_bellman_args *bellman_args,
-                               void (*bellman_fn)(long double *, int *,
-                                                  long double *, long double *, size_t)) {
+                               void (*bellman_fn)(uint64_t *, int *,
+                                                  uint64_t *, uint64_t *, int)) {
     static clock_t start, end;
     static double cpu_time_used;
